chore: rename navbar names to more descriptive
